// src/modules/satellites/types.ts
export interface SatellitePosition {
    lat: number;
    lon: number;
    alt: number;
    velocity: number;
}

export interface SatelliteInfo {
    id: string;
    name: string;
    category: string;
    noradId?: string;
}

export interface SimulatedSatellite extends SatelliteInfo {
    position: SatellitePosition;
    line1: string;
    line2: string;
    history: SatellitePosition[];
    orbitPath?: SatellitePosition[];
    isSelected: boolean;
    isHovered: boolean;
}